test: cover desktop updater safety paths
