+++ packages/petdex-desktop/sidecar/server.js
@@ -1,9 +1,9 @@
 #!/usr/bin/env node
-var h4=Object.create;var{getPrototypeOf:v4,defineProperty:W4,getOwnPropertyNames:U4}=Object;var P4=Object.prototype.hasOwnProperty;function g4($){return this[$]}var R4,A4,m4=($,W,J)=>{var Z=$!=null&&typeof $==="object";if(Z){var X=W?R4??=new WeakMap:A4??=new WeakMap,Q=X.get($);if(Q)return Q}J=$!=null?h4(v4($)):{};let Y=W||!$||!$.__esModule?W4(J,"default",{value:$,enumerable:!0}):J;for(let V of U4($))if(!P4.call(Y,V))W4(Y,V,{get:g4.bind($,V),enumerable:!0});if(Z)X.set($,Y);return Y};var s=require("node:child_process"),D=require("node:crypto"),G=require("node:fs"),a=m4(require("node:http")),C=require("node:os"),H=require("node:path");var J4="running-left";function Z4(){let $=J4;return J4=$==="running-left"?"running-right":"running-left",$}var T4={minDwellMs:250,maxQueueSize:50,durationStates:new Set(["waving","failed","review","jumping"])};class u{opts;pending=[];displayed=null;constructor($={}){this.opts={...T4,...$}}enqueue($){if(this.coalesces($))return!1;this.pending.push($);while(this.pending.length>this.opts.maxQueueSize)this.pending.shift();return!0}coalesces($){if(this.displayed&&this.displayed.state===$.state){if($.receivedAt-this.displayed.shownAt<this.displayed.duration)return!0}if(this.pending.length===0)return!1;if(this.pending[this.pending.length-1].state===$.state)return!0;return!1}tick($){if(this.displayed){if($-this.displayed.shownAt<this.displayed.duration)return null;this.displayed=null}let W=this.pending.shift();if(!W)return null;return this.displayed={state:W.state,duration:this.dwellFor(W),shownAt:$},W}dwellFor($){if(this.opts.durationStates.has($.state)&&$.duration)return Math.max($.duration,this.opts.minDwellMs);if($.duration&&$.duration>this.opts.minDwellMs)return $.duration;return this.opts.minDwellMs}pendingLength(){return this.pending.length}currentDisplayed(){return this.displayed?{...this.displayed}:null}}var q=require("node:path"),f={autoInstallUpdates:!0};function Q4($){try{let W=JSON.parse($);return{autoInstallUpdates:typeof W.autoInstallUpdates==="boolean"?W.autoInstallUpdates:f.autoInstallUpdates}}catch{return f}}function X4($){let W=q.resolve($);while(!0){if(q.basename(W).endsWith(".app"))return W;let J=q.dirname(W);if(J===W)return null;W=J}}function Y4($,W){let Z=`Petdex-${W==="arm64"?"arm64":W==="x64"?"x64":W}.dmg`;return $.assets.find((X)=>X.name===Z)??null}function G4($){let W=null;for(let J of $.split(`
-`)){let Z=J.indexOf("/Volumes/");if(Z===-1)continue;let X=J.slice(Z).trim();if(!W||X.length>W.length)W=X}return W}var B=Number(process.env.PETDEX_PORT??7777),x=H.join(C.homedir(),".petdex","runtime"),j4=H.join(x,"state.json"),x4=H.join(x,"bubble.json"),i=H.join(x,"update.json"),E4=H.join(x,"update.log"),V4=H.join(x,"update-token"),p=H.join(C.homedir(),".petdex","version"),z4=H.join(C.homedir(),".petdex","preferences.json"),d=H.join(x,"init-status.json"),S4=H.join(C.homedir(),".petdex","bin","petdex.js"),c4=H.join(x,"sidecar.log"),u4=65536,l4="https://api.github.com/repos/crafter-station/petdex/releases",K4=30,i4=5,p4="desktop-v",d4=21600000,o4=30000,m="x-petdex-update-token",M4=new Set(["idle","running","running-left","running-right","waving","jumping","failed","review","waiting"]);G.mkdirSync(x,{recursive:!0});var s4=30,O4=60,H4=(()=>{let $=O4,W=Date.now();return{consume(){let J=Date.now(),Z=(J-W)/1000;if(Z>0)$=Math.min(O4,$+Z*s4),W=J;if($<1)return!1;return $-=1,!0}}})(),v=D.randomBytes(32).toString("hex");function a4(){try{G.writeFileSync(V4,v,{mode:384}),G.chmodSync(V4,384)}catch($){process.stderr.write(`petdex sidecar: could not persist update token: ${$.message}
-`)}}var L4=H.join(C.homedir(),".petdex","telemetry.json"),n4=process.env.PETDEX_TELEMETRY_URL??"https://petdex.crafter.run/api/telemetry/event",k4=!1;function r4(){if(process.env.PETDEX_TELEMETRY==="0")return null;if(!G.existsSync(L4))return null;try{let $=JSON.parse(G.readFileSync(L4,"utf8"));if(typeof $.install_id!=="string")return null;if($.enabled===!1)return null;return{install_id:$.install_id,enabled:!0}}catch{return null}}function t4($,W){if(k4)return;k4=!0;let J=r4();if(!J)return;let Z=JSON.stringify({install_id:J.install_id,event:"desktop_first_state_received",state:$,agent_source:W});fetch(n4,{method:"POST",headers:{"content-type":"application/json"},body:Z,signal:AbortSignal.timeout(2000)}).catch(()=>{})}function T($,W){let J=Buffer.from($),Z=Buffer.from(W);if(J.length!==Z.length)return D.timingSafeEqual(J,J),!1;return D.timingSafeEqual(J,Z)}function z($){let W=`[${new Date().toISOString()}] ${$}
-`;try{G.appendFileSync(c4,W)}catch{}process.stderr.write(W)}var I=null,N4=0;function n($,W){N4+=1;let J={state:$,duration:W??null,updatedAt:Date.now(),counter:N4};if(G.writeFileSync(j4,JSON.stringify(J)),I)clearTimeout(I),I=null;if(typeof W==="number"&&W>0&&$!=="idle")I=setTimeout(()=>{n("idle"),I=null},W)}var r=new u({minDwellMs:250,maxQueueSize:50}),o=0;function B4($,W){o+=1;let J={text:$,agent_source:W,updatedAt:Date.now(),counter:o};G.writeFileSync(x4,JSON.stringify(J))}setInterval(()=>{let $=r.tick(Date.now());if($)n($.state,$.duration)},100).unref();n("idle");try{B4("",null)}catch{}try{let $=N(),W=A();if($.status==="available"&&($.current!==W||W===null))k({available:!1,current:W,latest:null,status:"idle",checkedAt:0}),z(`cleared stale update.json: persisted.current=${$.current??"?"} installed=${W??"?"}`)}catch{}S();function K($,W,J){let Z=JSON.stringify(J);$.writeHead(W,{"content-type":"application/json","content-length":Buffer.byteLength(Z)}),$.end(Z)}async function C4($){return new Promise((W,J)=>{let Z=[],X=0;$.on("data",(Q)=>{if(X+=Q.length,X>u4){$.destroy(Error("payload_too_large"));return}Z.push(Q)}),$.on("end",()=>{try{let Q=Buffer.concat(Z).toString("utf8");W(Q.length===0?{}:JSON.parse(Q))}catch(Q){J(Q)}}),$.on("error",J)})}function U(){return process.platform==="darwin"&&!!e()}function t($){return $?`Update ${$} is available. Run petdex update in your terminal.`:"Run petdex update in your terminal to update this install."}function A(){if(G.existsSync(p))try{let W=G.readFileSync(p,"utf8").trim();if(W)return W}catch{}let $=e();if(!$)return null;try{let J=G.readFileSync(H.join($,"Contents","Info.plist"),"utf8").match(/<key>CFBundleShortVersionString<\/key>\s*<string>([^<]+)<\/string>/);return J?.[1]?`desktop-v${J[1].trim()}`:null}catch{return null}}function e4(){if(!G.existsSync(z4))return f;try{return Q4(G.readFileSync(z4,"utf8"))}catch{return f}}function N(){if(!G.existsSync(i))return{available:!1,installable:U(),current:A(),latest:null,status:"idle",checkedAt:0};try{let $=JSON.parse(G.readFileSync(i,"utf8")),W=U();if($.available&&!W)return{...$,installable:!1,status:"available",message:$.message??t($.latest)};return{...$,installable:$.installable??W}}catch{return{available:!1,installable:U(),current:A(),latest:null,status:"idle",checkedAt:0}}}function k($){try{G.writeFileSync(i,JSON.stringify($))}catch(W){z(`update.json write failed: ${W.message}`)}if($.status==="error")try{r.enqueue({state:"failed",duration:3000,receivedAt:Date.now()})}catch(W){z(`mascot failed-state enqueue failed: ${W.message}`)}}function S(){let $=G.existsSync(S4);try{G.writeFileSync(d,JSON.stringify({needsInit:!$,reason:$?null:"no_hooks_installed",checkedAt:Date.now()}))}catch(W){z(`init-status.json write failed: ${W.message}`)}}async function q4(){for(let $=1;$<=i4;$++){let W=`${l4}?per_page=${K4}&page=${$}`,J=await fetch(W,{headers:{Accept:"application/vnd.github+json"},signal:AbortSignal.timeout(8000)});if(!J.ok)return z(`update check: GH API ${J.status} on page ${$}`),null;let Z=await J.json();if(!Array.isArray(Z)||Z.length===0)return null;let X=Z.find((Q)=>!Q.draft&&!Q.prerelease&&typeof Q.tag_name==="string"&&Q.tag_name.startsWith(p4));if(X?.tag_name)return{tag_name:X.tag_name,assets:Array.isArray(X.assets)?X.assets:[]};if(Z.length<K4)return null}return null}async function F4(){let $=A(),W=null;try{W=await q4()}catch(V){z(`update check failed: ${V.message}`);return}let J=W?.tag_name??null;if(N().status==="running")return;let X=!!J&&J!==$,Q=U(),Y={available:X,installable:Q,current:$,latest:J,status:X?"available":"idle",message:X&&!Q?t(J):void 0,checkedAt:Date.now()};if(k(Y),S(),z(`update check: current=${$??"?"} latest=${J??"?"} available=${X} installable=${Q}`),X&&Q&&e4().autoInstallUpdates)k({...Y,status:"running",message:"Installing the latest desktop release...",checkedAt:Date.now()}),b(`auto-install triggered for ${J}`),y4()}function b($){try{G.appendFileSync(E4,`[${new Date().toISOString()}] ${$}
-`)}catch{}}var P=null,g=!1,R=null,y=!1;function e(){return process.env.PETDEX_APP_BUNDLE??X4(process.argv[1]??"")}function $1($){return new Promise((W)=>setTimeout(W,$))}function W1($){try{return process.kill($,0),!0}catch(W){return W.code!=="ESRCH"}}async function F($,W,J={}){return new Promise((Z,X)=>{let Q=s.spawn($,W,{stdio:["ignore","pipe","pipe"]});P=Q;let Y="",V="";Q.stdout?.on("data",(M)=>{let O=M.toString("utf8");Y+=O;let L=O.trimEnd();if(L)b(`${$}: ${L}`)}),Q.stderr?.on("data",(M)=>{let O=M.toString("utf8");V+=O;let L=O.trimEnd();if(L)b(`${$} stderr: ${L}`)}),Q.on("error",(M)=>{P=null,X(M)}),Q.on("exit",(M)=>{if(P=null,M!==0&&!J.allowFailure){X(Error(`${$} exited with code ${M??"null"}: ${V||Y||"no output"}`));return}Z({stdout:Y,stderr:V,code:M})})})}function J1($){if(typeof $.digest!=="string")throw Error(`Release asset ${$.name??"unknown"} has no sha256 digest.`);if(!$.digest.startsWith("sha256:"))throw Error(`Release asset ${$.name??"unknown"} has an unsupported digest.`);return $.digest.slice(7).toLowerCase()}function D4($,W){return new Promise((J,Z)=>{let X=()=>{Y(),J()},Q=(V)=>{Y(),Z(V)},Y=()=>{$.off(W,X),$.off("error",Q)};$.once(W,X),$.once("error",Q)})}async function Z1($,W,J,Z){if(!Number.isSafeInteger(Z)||Z<=0)throw Error(`release asset has invalid size: ${Z}`);let X=`${W}.tmp`;try{let Q=await fetch($,{signal:AbortSignal.timeout(120000)});if(!Q.ok)throw Error(`download ${Q.status}`);if(!Q.body)throw Error("download response has no body");let Y=G.createWriteStream(X),V=Q.body.getReader(),M=D.createHash("sha256"),O=0,L=null;Y.on("error",(_)=>{L=_});try{while(!0){let{done:f4,value:I4}=await V.read();if(f4)break;let c=Buffer.from(I4);if(O+=c.byteLength,O>Z)throw Error(`download size mismatch: expected ${Z}, got more than ${O}`);if(M.update(c),!Y.write(c))await D4(Y,"drain");if(L)throw L}let _=D4(Y,"finish");Y.end(),await _}catch(_){throw Y.destroy(),_}finally{V.releaseLock()}if(O!==Z)throw Error(`download size mismatch: expected ${Z}, got ${O}`);let w=M.digest("hex");if(w!==J)throw Error(`download digest mismatch: expected ${J}, got ${w}`);G.renameSync(X,W)}catch(Q){throw G.rmSync(X,{force:!0}),Q}}async function Q1(){let $=Number(process.env.PETDEX_PARENT_PID);if(!Number.isFinite($)||$<=0)return;try{process.kill($,"SIGTERM")}catch(J){if(J.code!=="ESRCH")throw J;return}let W=Date.now()+1e4;while(Date.now()<W){if(!W1($))return;await $1(250)}throw Error("Petdex is still running after the quit request. Quit Petdex and try again.")}async function X1(){if(y)return;y=!0,await new Promise(($)=>j.close(()=>$()))}function Y1($){let W=[],J=null;for(let Z of $.split(`
-`)){if(Z.startsWith("Authority="))W.push(Z.slice(10).trim());if(Z.startsWith("TeamIdentifier=")){let X=Z.slice(15).trim();J=X&&X!=="not set"?X:null}}return{teamIdentifier:J,authorities:W}}async function b4($){let W=await F("codesign",["-dv","--verbose=4",$]);return Y1(`${W.stdout}
-${W.stderr}`)}async function l($,W){await F("codesign",["--verify","--deep","--strict",$]);let J=await b4($);G1(W,J),await F("spctl",["-a","-t","exec",$])}function G1($,W){if(!$.teamIdentifier)throw Error("Current Petdex.app has no Developer ID team identifier.");if(!W.teamIdentifier)throw Error("Update Petdex.app has no Developer ID team identifier.");if($.teamIdentifier!==W.teamIdentifier)throw Error(`Update signer mismatch: expected ${$.teamIdentifier}, got ${W.teamIdentifier}.`);if(!W.authorities.some((J)=>J.startsWith("Developer ID Application:")))throw Error("Update Petdex.app is not signed with Developer ID Application.")}function V1($){let W=D.randomBytes(8).toString("hex"),J=H.dirname($),Z=H.basename($);return{stagedApp:H.join(J,`.${Z}.update-${W}`),backupApp:H.join(J,`.${Z}.previous-${W}`)}}async function z1($,W,J,Z){G.rmSync(J,{recursive:!0,force:!0}),G.renameSync($,J);let X=!1;try{G.renameSync(W,$),X=!0,await Z(),G.rmSync(J,{recursive:!0,force:!0})}catch(Q){if(X)G.rmSync($,{recursive:!0,force:!0});if(G.existsSync(J))G.renameSync(J,$);throw Q}}async function K1(){if(process.platform!=="darwin")throw Error("Bundled desktop updater is only available on macOS.");let $=await q4();if(!$)throw Error("No desktop release found.");let W=A();if(W===$.tag_name){k({available:!1,current:W,latest:$.tag_name,status:"done",message:"Already up to date.",checkedAt:Date.now()});return}let J=Y4($,C.arch());if(!J)throw Error(`No DMG asset for ${C.arch()} in ${$.tag_name}.`);let Z=e();if(!Z)throw Error("Petdex.app bundle not found for in-app update.");let X=H.join(C.tmpdir(),`petdex-${D.randomBytes(8).toString("hex")}-${J.name}`),Q=null,Y=null,V=null;try{k({...N(),latest:$.tag_name,status:"running",message:`Downloading ${$.tag_name}...`,checkedAt:Date.now()}),await Z1(J.browser_download_url,X,J1(J),J.size),k({...N(),status:"running",message:"Mounting update image...",checkedAt:Date.now()});let M=await F("hdiutil",["attach","-nobrowse",X]);if(Q=G4(M.stdout),!Q)throw Error("Could not determine mounted update volume.");let O=H.join(Q,"Petdex.app");if(!G.existsSync(O))throw Error(`Mounted update does not contain ${O}.`);k({...N(),status:"running",message:"Verifying update signature...",checkedAt:Date.now()}),await F("codesign",["--verify","--deep","--strict",O]),await F("codesign",["--verify","--deep","--strict",Z]);let L=await b4(Z);await l(O,L),k({...N(),status:"running",message:"Replacing Petdex.app...",checkedAt:Date.now()});let w=V1(Z);Y=w.stagedApp,G.rmSync(Y,{recursive:!0,force:!0}),await F("ditto",[O,Y]),await l(Y,L),await Q1(),await z1(Z,Y,w.backupApp,async()=>{await F("xattr",["-dr","com.apple.quarantine",Z],{allowFailure:!0}),await l(Z,L)}),G.writeFileSync(p,`${$.tag_name}
-`),k({available:!1,current:$.tag_name,latest:$.tag_name,status:"done",message:"Update installed. Relaunching Petdex.",checkedAt:Date.now()}),b(`installed ${$.tag_name} into ${Z}`),V=Z}finally{if(Q)await F("hdiutil",["detach","-quiet",Q],{allowFailure:!0}).catch(()=>{});if(G.rmSync(X,{force:!0}),Y)G.rmSync(Y,{recursive:!0,force:!0})}if(V)await X1(),s.spawn("open",[V],{detached:!0,stdio:"ignore"}).unref(),setTimeout(()=>process.exit(0),250).unref()}function y4(){if(g)return;g=!0,R=K1().catch(($)=>{let W=$.message;z(`spawnUpdate: ${W}`),b(`error: ${W}`),k({...N(),status:"error",message:W,checkedAt:Date.now()})}).finally(()=>{g=!1,P=null,R=null})}var j=a.default.createServer(async($,W)=>{try{let J=new URL($.url??"/",`http://127.0.0.1:${B}`);if($.method==="GET"&&J.pathname==="/health")return K(W,200,{ok:!0,port:B});if($.method==="GET"&&J.pathname==="/whoami"){let Z=Number(process.env.PETDEX_PARENT_PID);return K(W,200,{ok:!0,pid:process.pid,parentPid:Number.isFinite(Z)&&Z>0?Z:null})}if($.method==="GET"&&J.pathname==="/state")try{let{readFileSync:Z}=await import("node:fs"),X=Z(j4,"utf8");W.writeHead(200,{"content-type":"application/json"}),W.end(X);return}catch{return K(W,200,{state:"idle",counter:0})}if($.method==="POST"&&J.pathname==="/state"){let Z=$.headers[m],X=Array.isArray(Z)?Z[0]:Z;if(!X||!T(X,v))return K(W,401,{ok:!1,error:"unauthorized"});if(!H4.consume())return K(W,429,{ok:!1,error:"rate_limited"});let Q;try{Q=await C4($)}catch{return K(W,400,{ok:!1,error:"invalid_json"})}let Y=Q,V=typeof Y.state==="string"?Y.state:null;if(!V||!M4.has(V))return K(W,400,{ok:!1,error:"invalid_state",valid:[...M4]});let M=typeof Y.duration==="number"&&Y.duration>0?Math.min(Y.duration,30000):void 0,O=typeof Y.agent_source==="string"?Y.agent_source.slice(0,64):null,L=V==="running"?Z4():V,w=r.enqueue({state:L,duration:M,receivedAt:Date.now()});return z(`state=${L} duration=${M??"-"} ${w?"queued":"coalesced"}`),S(),t4(V,O),K(W,200,{ok:!0,state:V,duration:M??null,queued:w})}if($.method==="GET"&&J.pathname==="/bubble")try{let{readFileSync:Z}=await import("node:fs"),X=Z(x4,"utf8");W.writeHead(200,{"content-type":"application/json"}),W.end(X);return}catch{return K(W,200,{text:"",counter:0})}if($.method==="POST"&&J.pathname==="/bubble"){let Z=$.headers[m],X=Array.isArray(Z)?Z[0]:Z;if(!X||!T(X,v))return K(W,401,{ok:!1,error:"unauthorized"});if(!H4.consume())return K(W,429,{ok:!1,error:"rate_limited"});let Q;try{Q=await C4($)}catch{return K(W,400,{ok:!1,error:"invalid_json"})}let Y=Q,V=typeof Y.text==="string"?Y.text:null;if(!V)return K(W,400,{ok:!1,error:"missing_text"});let M=V.slice(0,200),O=typeof Y.agent_source==="string"?Y.agent_source.slice(0,64):null;return B4(M,O),S(),z(`bubble="${M.slice(0,60)}" source=${O??"-"}`),K(W,200,{ok:!0,text:M,counter:o})}if($.method==="GET"&&J.pathname==="/update")return K(W,200,N());if($.method==="GET"&&J.pathname==="/init-status"){if(!G.existsSync(d))return K(W,404,{needsInit:!0,reason:"no_hooks_installed",checkedAt:0});try{let Z=G.readFileSync(d,"utf8");W.writeHead(200,{"content-type":"application/json"}),W.end(Z);return}catch{return K(W,200,{needsInit:!1,reason:null,checkedAt:0})}}if($.method==="POST"&&J.pathname==="/update/handoff"){let Z=$.headers[m],X=Array.isArray(Z)?Z[0]:Z;if(!X||!T(X,v))return K(W,401,{ok:!1,error:"unauthorized"});if(!y){y=!0,z("handoff requested by updater; releasing port");let Q=N();if(Q.status==="running")k({...Q,status:"done",message:"Update installed. Restarting the desktop.",checkedAt:Date.now()});W.writeHead(200,{"content-type":"application/json"}),W.end(JSON.stringify({ok:!0})),j.close(()=>{z("handoff: server closed, exiting"),process.exit(0)}),setTimeout(()=>process.exit(0),1000).unref();return}return K(W,200,{ok:!0,alreadyHandedOff:!0})}if($.method==="POST"&&J.pathname==="/update"){let Z=$.headers[m],X=Array.isArray(Z)?Z[0]:Z;if(!X||!T(X,v))return K(W,401,{ok:!1,error:"unauthorized"});let Q=N();if(Q.status==="running")return K(W,200,Q);if(!Q.available&&Q.status!=="error")return K(W,200,Q);if(!U()){let V={...Q,available:!!Q.latest&&Q.latest!==Q.current,installable:!1,status:"available",message:t(Q.latest),checkedAt:Date.now()};return k(V),K(W,409,{ok:!1,error:"unsupported_install",message:V.message})}let Y={...Q,installable:!0,status:"running",message:"Installing the latest desktop release...",checkedAt:Date.now()};return k(Y),b("triggered by webview click"),y4(),K(W,202,Y)}K(W,404,{ok:!1,error:"not_found"})}catch(J){z(`server error: ${J.message}`),K(W,500,{ok:!1,error:"internal"})}}),M1=2,E=0;function _4(){E+=1,j.listen(B,"127.0.0.1")}j.on("listening",()=>{a4(),z(`petdex sidecar listening on http://127.0.0.1:${B}`)});j.on("error",($)=>{if($.code!=="EADDRINUSE"){z(`server.error: ${$.message}`),process.exit(1);return}O1($)});async function O1($){if(E>=M1){z(`server.error: ${$.message}; gave up after ${E} attempts`),process.exit(1);return}let W=null;try{let Q=await fetch(`http://127.0.0.1:${B}/whoami`,{signal:AbortSignal.timeout(1000)});if(Q.ok)W=await Q.json()}catch{z(`server.error: ${$.message}; /whoami probe failed, yielding`),process.exit(1);return}if(!W||typeof W.pid!=="number"){z(`server.error: ${$.message}; /whoami had no pid, yielding`),process.exit(1);return}let J=W.parentPid,Z=!0;if(typeof J==="number"&&J>0)try{process.kill(J,0)}catch(Q){Z=Q.code!=="ESRCH"}else{z(`server.error: ${$.message}; incumbent pid=${W.pid} has no parentPid, yielding`),process.exit(1);return}if(Z){z(`server.error: ${$.message}; incumbent pid=${W.pid} parent=${J} is alive, yielding`),process.exit(1);return}z(`server.error: ${$.message}; killing orphan sidecar pid=${W.pid} (parent=${J} dead)`);try{process.kill(W.pid,"SIGTERM")}catch(Q){z(`failed to SIGTERM orphan: ${Q.message}`),process.exit(1);return}if(!await H1(B,5000)){z(`port :${B} still busy 5s after killing orphan, giving up`),process.exit(1);return}z(`orphan evicted, retrying listen (attempt ${E+1})`),_4()}async function H1($,W){let J=Date.now()+W;while(Date.now()<J){if(await new Promise((X)=>{let Q=a.default.createServer().once("error",()=>X(!1)).once("listening",()=>Q.close(()=>X(!0)));Q.listen($,"127.0.0.1")}))return!0;await new Promise((X)=>setTimeout(X,200))}return!1}_4();var w4=60000;function $4($){if(y){z(`shutdown(${$}) ignored: handoff in progress`);return}if(g){z(`sidecar received ${$}; update in progress`);let W=Date.now(),J=setTimeout(()=>{z(`update still running after ${w4}ms; forcing exit`);let Z=N();if(Z.status==="running")k({...Z,status:"error",message:"Sidecar shut down before update finished. Re-launch Petdex and try again.",checkedAt:Date.now()});j.close(()=>process.exit(0)),setTimeout(()=>process.exit(0),1000).unref()},w4);if(!R){J.unref();return}R.finally(()=>{if(clearTimeout(J),y)return;z(`update settled after ${Date.now()-W}ms; shutting down`),j.close(()=>process.exit(0)),setTimeout(()=>process.exit(0),1000).unref()});return}z(`sidecar received ${$}, shutting down`),j.close(()=>process.exit(0)),setTimeout(()=>process.exit(0),1000).unref()}process.on("SIGTERM",()=>$4("SIGTERM"));process.on("SIGINT",()=>$4("SIGINT"));var L1=setTimeout(()=>{F4(),setInterval(()=>void F4(),d4).unref()},o4);L1.unref();var h=Number(process.env.PETDEX_PARENT_PID);if(Number.isFinite(h)&&h>0){z(`sidecar watching parent pid ${h}`);let $=setInterval(()=>{try{process.kill(h,0)}catch{if(g||P)return;z(`parent ${h} gone, exiting`),clearInterval($),$4("parent-gone")}},2000);$.unref()}
+var I$=Object.create;var{getPrototypeOf:c$,defineProperty:J$,getOwnPropertyNames:U$}=Object;var A$=Object.prototype.hasOwnProperty;function R$($){return this[$]}var S$,E$,T$=($,W,J)=>{var z=$!=null&&typeof $==="object";if(z){var Q=W?S$??=new WeakMap:E$??=new WeakMap,Z=Q.get($);if(Z)return Z}J=$!=null?I$(c$($)):{};let X=W||!$||!$.__esModule?J$(J,"default",{value:$,enumerable:!0}):J;for(let G of U$($))if(!A$.call(X,G))J$(X,G,{get:R$.bind($,G),enumerable:!0});if(z)Q.set($,X);return X};var n=require("node:child_process"),b=require("node:crypto"),Y=require("node:fs"),a=T$(require("node:http")),F=require("node:os"),M=require("node:path");var z$="running-left";function Z$(){let $=z$;return z$=$==="running-left"?"running-right":"running-left",$}var u$={minDwellMs:250,maxQueueSize:50,durationStates:new Set(["waving","failed","review","jumping"])};class l{opts;pending=[];displayed=null;constructor($={}){this.opts={...u$,...$}}enqueue($){if(this.coalesces($))return!1;this.pending.push($);while(this.pending.length>this.opts.maxQueueSize)this.pending.shift();return!0}coalesces($){if(this.displayed&&this.displayed.state===$.state){if($.receivedAt-this.displayed.shownAt<this.displayed.duration)return!0}if(this.pending.length===0)return!1;if(this.pending[this.pending.length-1].state===$.state)return!0;return!1}tick($){if(this.displayed){if($-this.displayed.shownAt<this.displayed.duration)return null;this.displayed=null}let W=this.pending.shift();if(!W)return null;return this.displayed={state:W.state,duration:this.dwellFor(W),shownAt:$},W}dwellFor($){if(this.opts.durationStates.has($.state)&&$.duration)return Math.max($.duration,this.opts.minDwellMs);if($.duration&&$.duration>this.opts.minDwellMs)return $.duration;return this.opts.minDwellMs}pendingLength(){return this.pending.length}currentDisplayed(){return this.displayed?{...this.displayed}:null}}var X$=require("node:crypto"),O=require("node:fs"),h=require("node:path"),v={autoInstallUpdates:!0};function Y$($){try{let W=JSON.parse($);return{autoInstallUpdates:typeof W.autoInstallUpdates==="boolean"?W.autoInstallUpdates:v.autoInstallUpdates}}catch{return v}}function G$($){let W=h.resolve($);while(!0){if(h.basename(W).endsWith(".app"))return W;let J=h.dirname(W);if(J===W)return null;W=J}}function V$($,W){let z=`Petdex-${W==="arm64"?"arm64":W==="x64"?"x64":W}.dmg`;return $.assets.find((Q)=>Q.name===z)??null}function w$($){let W=null;for(let J of $.split(`
+`)){let z=J.indexOf("/Volumes/");if(z===-1)continue;let Q=J.slice(z).trim();if(!W||Q.length>W.length)W=Q}return W}function Q$($,W){return new Promise((J,z)=>{let Q=()=>{X(),J()},Z=(G)=>{X(),z(G)},X=()=>{$.off(W,Q),$.off("error",Z)};$.once(W,Q),$.once("error",Z)})}async function K$($,W,J,z){if(!Number.isSafeInteger(z)||z<=0)throw Error(`release asset has invalid size: ${z}`);let Q=`${W}.tmp`;try{let Z=await fetch($,{signal:AbortSignal.timeout(120000)});if(!Z.ok)throw Error(`download ${Z.status}`);if(!Z.body)throw Error("download response has no body");let X=O.createWriteStream(Q),G=Z.body.getReader(),K=X$.createHash("sha256"),k=0,H=null;X.on("error",(q)=>{H=q});try{while(!0){let{done:P$,value:_$}=await G.read();if(P$)break;let u=Buffer.from(_$);if(k+=u.byteLength,k>z)throw Error(`download size mismatch: expected ${z}, got more than ${k}`);if(K.update(u),!X.write(u))await Q$(X,"drain");if(H)throw H}let q=Q$(X,"finish");X.end(),await q}catch(q){throw X.destroy(),q}finally{G.releaseLock()}if(k!==z)throw Error(`download size mismatch: expected ${z}, got ${k}`);let N=K.digest("hex");if(N!==J)throw Error(`download digest mismatch: expected ${J}, got ${N}`);O.renameSync(Q,W)}catch(Z){throw O.rmSync(Q,{force:!0}),Z}}async function k$($,W,J,z){O.rmSync(J,{recursive:!0,force:!0}),O.renameSync($,J);let Q=!1;try{O.renameSync(W,$),Q=!0,await z(),O.rmSync(J,{recursive:!0,force:!0})}catch(Z){if(Q)O.rmSync($,{recursive:!0,force:!0});if(O.existsSync(J))O.renameSync(J,$);throw Z}}var D=Number(process.env.PETDEX_PORT??7777),C=M.join(F.homedir(),".petdex","runtime"),h$=M.join(C,"state.json"),B$=M.join(C,"bubble.json"),p=M.join(C,"update.json"),l$=M.join(C,"update.log"),M$=M.join(C,"update-token"),o=M.join(F.homedir(),".petdex","version"),H$=M.join(F.homedir(),".petdex","preferences.json"),d=M.join(C,"init-status.json"),i$=M.join(F.homedir(),".petdex","bin","petdex.js"),p$=M.join(C,"sidecar.log"),o$=65536,d$="https://api.github.com/repos/crafter-station/petdex/releases",O$=30,s$=5,n$="desktop-v",a$=21600000,r$=30000,R="x-petdex-update-token",L$=new Set(["idle","running","running-left","running-right","waving","jumping","failed","review","waiting"]);Y.mkdirSync(C,{recursive:!0});var t$=30,y$=60,F$=(()=>{let $=y$,W=Date.now();return{consume(){let J=Date.now(),z=(J-W)/1000;if(z>0)$=Math.min(y$,$+z*t$),W=J;if($<1)return!1;return $-=1,!0}}})(),P=b.randomBytes(32).toString("hex");function e$(){try{Y.writeFileSync(M$,P,{mode:384}),Y.chmodSync(M$,384)}catch($){process.stderr.write(`petdex sidecar: could not persist update token: ${$.message}
+`)}}var x$=M.join(F.homedir(),".petdex","telemetry.json"),$0=process.env.PETDEX_TELEMETRY_URL??"https://petdex.crafter.run/api/telemetry/event",N$=!1;function W0(){if(process.env.PETDEX_TELEMETRY==="0")return null;if(!Y.existsSync(x$))return null;try{let $=JSON.parse(Y.readFileSync(x$,"utf8"));if(typeof $.install_id!=="string")return null;if($.enabled===!1)return null;return{install_id:$.install_id,enabled:!0}}catch{return null}}function J0($,W){if(N$)return;N$=!0;let J=W0();if(!J)return;let z=JSON.stringify({install_id:J.install_id,event:"desktop_first_state_received",state:$,agent_source:W});fetch($0,{method:"POST",headers:{"content-type":"application/json"},body:z,signal:AbortSignal.timeout(2000)}).catch(()=>{})}function S($,W){let J=Buffer.from($),z=Buffer.from(W);if(J.length!==z.length)return b.timingSafeEqual(J,J),!1;return b.timingSafeEqual(J,z)}function V($){let W=`[${new Date().toISOString()}] ${$}
+`;try{Y.appendFileSync(p$,W)}catch{}process.stderr.write(W)}var g=null,f$=0;function r($,W){f$+=1;let J={state:$,duration:W??null,updatedAt:Date.now(),counter:f$};if(Y.writeFileSync(h$,JSON.stringify(J)),g)clearTimeout(g),g=null;if(typeof W==="number"&&W>0&&$!=="idle")g=setTimeout(()=>{r("idle"),g=null},W)}var t=new l({minDwellMs:250,maxQueueSize:50}),s=0;function j$($,W){s+=1;let J={text:$,agent_source:W,updatedAt:Date.now(),counter:s};Y.writeFileSync(B$,JSON.stringify(J))}setInterval(()=>{let $=t.tick(Date.now());if($)r($.state,$.duration)},100).unref();r("idle");try{j$("",null)}catch{}try{let $=y(),W=A();if($.status==="available"&&($.current!==W||W===null))L({available:!1,current:W,latest:null,status:"idle",checkedAt:0}),V(`cleared stale update.json: persisted.current=${$.current??"?"} installed=${W??"?"}`)}catch{}T();function w($,W,J){let z=JSON.stringify(J);$.writeHead(W,{"content-type":"application/json","content-length":Buffer.byteLength(z)}),$.end(z)}async function C$($){return new Promise((W,J)=>{let z=[],Q=0;$.on("data",(Z)=>{if(Q+=Z.length,Q>o$){$.destroy(Error("payload_too_large"));return}z.push(Z)}),$.on("end",()=>{try{let Z=Buffer.concat(z).toString("utf8");W(Z.length===0?{}:JSON.parse(Z))}catch(Z){J(Z)}}),$.on("error",J)})}function _(){return process.platform==="darwin"&&!!$$()}function e($){return $?`Update ${$} is available. Run petdex update in your terminal.`:"Run petdex update in your terminal to update this install."}function A(){if(Y.existsSync(o))try{let W=Y.readFileSync(o,"utf8").trim();if(W)return W}catch{}let $=$$();if(!$)return null;try{let J=Y.readFileSync(M.join($,"Contents","Info.plist"),"utf8").match(/<key>CFBundleShortVersionString<\/key>\s*<string>([^<]+)<\/string>/);return J?.[1]?`desktop-v${J[1].trim()}`:null}catch{return null}}function z0(){if(!Y.existsSync(H$))return v;try{return Y$(Y.readFileSync(H$,"utf8"))}catch{return v}}function y(){if(!Y.existsSync(p))return{available:!1,installable:_(),current:A(),latest:null,status:"idle",checkedAt:0};try{let $=JSON.parse(Y.readFileSync(p,"utf8")),W=_();if($.available&&!W)return{...$,installable:!1,status:"available",message:$.message??e($.latest)};return{...$,installable:$.installable??W}}catch{return{available:!1,installable:_(),current:A(),latest:null,status:"idle",checkedAt:0}}}function L($){try{Y.writeFileSync(p,JSON.stringify($))}catch(W){V(`update.json write failed: ${W.message}`)}if($.status==="error")try{t.enqueue({state:"failed",duration:3000,receivedAt:Date.now()})}catch(W){V(`mascot failed-state enqueue failed: ${W.message}`)}}function T(){let $=Y.existsSync(i$);try{Y.writeFileSync(d,JSON.stringify({needsInit:!$,reason:$?null:"no_hooks_installed",checkedAt:Date.now()}))}catch(W){V(`init-status.json write failed: ${W.message}`)}}async function q$(){for(let $=1;$<=s$;$++){let W=`${d$}?per_page=${O$}&page=${$}`,J=await fetch(W,{headers:{Accept:"application/vnd.github+json"},signal:AbortSignal.timeout(8000)});if(!J.ok)return V(`update check: GH API ${J.status} on page ${$}`),null;let z=await J.json();if(!Array.isArray(z)||z.length===0)return null;let Q=z.find((Z)=>!Z.draft&&!Z.prerelease&&typeof Z.tag_name==="string"&&Z.tag_name.startsWith(n$));if(Q?.tag_name)return{tag_name:Q.tag_name,assets:Array.isArray(Q.assets)?Q.assets:[]};if(z.length<O$)return null}return null}async function D$(){let $=A(),W=null;try{W=await q$()}catch(G){V(`update check failed: ${G.message}`);return}let J=W?.tag_name??null;if(y().status==="running")return;let Q=!!J&&J!==$,Z=_(),X={available:Q,installable:Z,current:$,latest:J,status:Q?"available":"idle",message:Q&&!Z?e(J):void 0,checkedAt:Date.now()};if(L(X),T(),V(`update check: current=${$??"?"} latest=${J??"?"} available=${Q} installable=${Z}`),Q&&Z&&z0().autoInstallUpdates)L({...X,status:"running",message:"Installing the latest desktop release...",checkedAt:Date.now()}),B(`auto-install triggered for ${J}`),g$()}function B($){try{Y.appendFileSync(l$,`[${new Date().toISOString()}] ${$}
+`)}catch{}}var I=null,c=!1,U=null,j=!1;function $$(){return process.env.PETDEX_APP_BUNDLE??G$(process.argv[1]??"")}function Z0($){return new Promise((W)=>setTimeout(W,$))}function Q0($){try{return process.kill($,0),!0}catch(W){return W.code!=="ESRCH"}}async function x($,W,J={}){return new Promise((z,Q)=>{let Z=n.spawn($,W,{stdio:["ignore","pipe","pipe"]});I=Z;let X="",G="";Z.stdout?.on("data",(K)=>{let k=K.toString("utf8");X+=k;let H=k.trimEnd();if(H)B(`${$}: ${H}`)}),Z.stderr?.on("data",(K)=>{let k=K.toString("utf8");G+=k;let H=k.trimEnd();if(H)B(`${$} stderr: ${H}`)}),Z.on("error",(K)=>{I=null,Q(K)}),Z.on("exit",(K)=>{if(I=null,K!==0&&!J.allowFailure){Q(Error(`${$} exited with code ${K??"null"}: ${G||X||"no output"}`));return}z({stdout:X,stderr:G,code:K})})})}function X0($){if(typeof $.digest!=="string")throw Error(`Release asset ${$.name??"unknown"} has no sha256 digest.`);if(!$.digest.startsWith("sha256:"))throw Error(`Release asset ${$.name??"unknown"} has an unsupported digest.`);return $.digest.slice(7).toLowerCase()}async function Y0(){let $=Number(process.env.PETDEX_PARENT_PID);if(!Number.isFinite($)||$<=0)return;try{process.kill($,"SIGTERM")}catch(J){if(J.code!=="ESRCH")throw J;return}let W=Date.now()+1e4;while(Date.now()<W){if(!Q0($))return;await Z0(250)}throw Error("Petdex is still running after the quit request. Quit Petdex and try again.")}async function G0(){if(j)return;j=!0,await new Promise(($)=>f.close(()=>$()))}function V0($){let W=[],J=null;for(let z of $.split(`
+`)){if(z.startsWith("Authority="))W.push(z.slice(10).trim());if(z.startsWith("TeamIdentifier=")){let Q=z.slice(15).trim();J=Q&&Q!=="not set"?Q:null}}return{teamIdentifier:J,authorities:W}}async function v$($){let W=await x("codesign",["-dv","--verbose=4",$]);return V0(`${W.stdout}
+${W.stderr}`)}async function i($,W){await x("codesign",["--verify","--deep","--strict",$]);let J=await v$($);w0(W,J),await x("spctl",["-a","-t","exec",$])}function w0($,W){if(!$.teamIdentifier)throw Error("Current Petdex.app has no Developer ID team identifier.");if(!W.teamIdentifier)throw Error("Update Petdex.app has no Developer ID team identifier.");if($.teamIdentifier!==W.teamIdentifier)throw Error(`Update signer mismatch: expected ${$.teamIdentifier}, got ${W.teamIdentifier}.`);if(!W.authorities.some((J)=>J.startsWith("Developer ID Application:")))throw Error("Update Petdex.app is not signed with Developer ID Application.")}function K0($){let W=b.randomBytes(8).toString("hex"),J=M.dirname($),z=M.basename($);return{stagedApp:M.join(J,`.${z}.update-${W}`),backupApp:M.join(J,`.${z}.previous-${W}`)}}async function k0(){if(process.platform!=="darwin")throw Error("Bundled desktop updater is only available on macOS.");let $=await q$();if(!$)throw Error("No desktop release found.");let W=A();if(W===$.tag_name){L({available:!1,current:W,latest:$.tag_name,status:"done",message:"Already up to date.",checkedAt:Date.now()});return}let J=V$($,F.arch());if(!J)throw Error(`No DMG asset for ${F.arch()} in ${$.tag_name}.`);let z=$$();if(!z)throw Error("Petdex.app bundle not found for in-app update.");let Q=M.join(F.tmpdir(),`petdex-${b.randomBytes(8).toString("hex")}-${J.name}`),Z=null,X=null,G=null;try{L({...y(),latest:$.tag_name,status:"running",message:`Downloading ${$.tag_name}...`,checkedAt:Date.now()}),await K$(J.browser_download_url,Q,X0(J),J.size),L({...y(),status:"running",message:"Mounting update image...",checkedAt:Date.now()});let K=await x("hdiutil",["attach","-nobrowse",Q]);if(Z=w$(K.stdout),!Z)throw Error("Could not determine mounted update volume.");let k=M.join(Z,"Petdex.app");if(!Y.existsSync(k))throw Error(`Mounted update does not contain ${k}.`);L({...y(),status:"running",message:"Verifying update signature...",checkedAt:Date.now()}),await x("codesign",["--verify","--deep","--strict",k]),await x("codesign",["--verify","--deep","--strict",z]);let H=await v$(z);await i(k,H),L({...y(),status:"running",message:"Replacing Petdex.app...",checkedAt:Date.now()});let N=K0(z);X=N.stagedApp,Y.rmSync(X,{recursive:!0,force:!0}),await x("ditto",[k,X]),await i(X,H),await Y0(),await k$(z,X,N.backupApp,async()=>{await x("xattr",["-dr","com.apple.quarantine",z],{allowFailure:!0}),await i(z,H)}),Y.writeFileSync(o,`${$.tag_name}
+`),L({available:!1,current:$.tag_name,latest:$.tag_name,status:"done",message:"Update installed. Relaunching Petdex.",checkedAt:Date.now()}),B(`installed ${$.tag_name} into ${z}`),G=z}finally{if(Z)await x("hdiutil",["detach","-quiet",Z],{allowFailure:!0}).catch(()=>{});if(Y.rmSync(Q,{force:!0}),X)Y.rmSync(X,{recursive:!0,force:!0})}if(G)await G0(),n.spawn("open",[G],{detached:!0,stdio:"ignore"}).unref(),setTimeout(()=>process.exit(0),250).unref()}function g$(){if(c)return;c=!0,U=k0().catch(($)=>{let W=$.message;V(`spawnUpdate: ${W}`),B(`error: ${W}`),L({...y(),status:"error",message:W,checkedAt:Date.now()})}).finally(()=>{c=!1,I=null,U=null})}var f=a.default.createServer(async($,W)=>{try{let J=new URL($.url??"/",`http://127.0.0.1:${D}`);if($.method==="GET"&&J.pathname==="/health")return w(W,200,{ok:!0,port:D});if($.method==="GET"&&J.pathname==="/whoami"){let z=Number(process.env.PETDEX_PARENT_PID);return w(W,200,{ok:!0,pid:process.pid,parentPid:Number.isFinite(z)&&z>0?z:null})}if($.method==="GET"&&J.pathname==="/state")try{let{readFileSync:z}=await import("node:fs"),Q=z(h$,"utf8");W.writeHead(200,{"content-type":"application/json"}),W.end(Q);return}catch{return w(W,200,{state:"idle",counter:0})}if($.method==="POST"&&J.pathname==="/state"){let z=$.headers[R],Q=Array.isArray(z)?z[0]:z;if(!Q||!S(Q,P))return w(W,401,{ok:!1,error:"unauthorized"});if(!F$.consume())return w(W,429,{ok:!1,error:"rate_limited"});let Z;try{Z=await C$($)}catch{return w(W,400,{ok:!1,error:"invalid_json"})}let X=Z,G=typeof X.state==="string"?X.state:null;if(!G||!L$.has(G))return w(W,400,{ok:!1,error:"invalid_state",valid:[...L$]});let K=typeof X.duration==="number"&&X.duration>0?Math.min(X.duration,30000):void 0,k=typeof X.agent_source==="string"?X.agent_source.slice(0,64):null,H=G==="running"?Z$():G,N=t.enqueue({state:H,duration:K,receivedAt:Date.now()});return V(`state=${H} duration=${K??"-"} ${N?"queued":"coalesced"}`),T(),J0(G,k),w(W,200,{ok:!0,state:G,duration:K??null,queued:N})}if($.method==="GET"&&J.pathname==="/bubble")try{let{readFileSync:z}=await import("node:fs"),Q=z(B$,"utf8");W.writeHead(200,{"content-type":"application/json"}),W.end(Q);return}catch{return w(W,200,{text:"",counter:0})}if($.method==="POST"&&J.pathname==="/bubble"){let z=$.headers[R],Q=Array.isArray(z)?z[0]:z;if(!Q||!S(Q,P))return w(W,401,{ok:!1,error:"unauthorized"});if(!F$.consume())return w(W,429,{ok:!1,error:"rate_limited"});let Z;try{Z=await C$($)}catch{return w(W,400,{ok:!1,error:"invalid_json"})}let X=Z,G=typeof X.text==="string"?X.text:null;if(!G)return w(W,400,{ok:!1,error:"missing_text"});let K=G.slice(0,200),k=typeof X.agent_source==="string"?X.agent_source.slice(0,64):null;return j$(K,k),T(),V(`bubble="${K.slice(0,60)}" source=${k??"-"}`),w(W,200,{ok:!0,text:K,counter:s})}if($.method==="GET"&&J.pathname==="/update")return w(W,200,y());if($.method==="GET"&&J.pathname==="/init-status"){if(!Y.existsSync(d))return w(W,404,{needsInit:!0,reason:"no_hooks_installed",checkedAt:0});try{let z=Y.readFileSync(d,"utf8");W.writeHead(200,{"content-type":"application/json"}),W.end(z);return}catch{return w(W,200,{needsInit:!1,reason:null,checkedAt:0})}}if($.method==="POST"&&J.pathname==="/update/handoff"){let z=$.headers[R],Q=Array.isArray(z)?z[0]:z;if(!Q||!S(Q,P))return w(W,401,{ok:!1,error:"unauthorized"});if(!j){j=!0,V("handoff requested by updater; releasing port");let Z=y();if(Z.status==="running")L({...Z,status:"done",message:"Update installed. Restarting the desktop.",checkedAt:Date.now()});W.writeHead(200,{"content-type":"application/json"}),W.end(JSON.stringify({ok:!0})),f.close(()=>{V("handoff: server closed, exiting"),process.exit(0)}),setTimeout(()=>process.exit(0),1000).unref();return}return w(W,200,{ok:!0,alreadyHandedOff:!0})}if($.method==="POST"&&J.pathname==="/update"){let z=$.headers[R],Q=Array.isArray(z)?z[0]:z;if(!Q||!S(Q,P))return w(W,401,{ok:!1,error:"unauthorized"});let Z=y();if(Z.status==="running")return w(W,200,Z);if(!Z.available&&Z.status!=="error")return w(W,200,Z);if(!_()){let G={...Z,available:!!Z.latest&&Z.latest!==Z.current,installable:!1,status:"available",message:e(Z.latest),checkedAt:Date.now()};return L(G),w(W,409,{ok:!1,error:"unsupported_install",message:G.message})}let X={...Z,installable:!0,status:"running",message:"Installing the latest desktop release...",checkedAt:Date.now()};return L(X),B("triggered by webview click"),g$(),w(W,202,X)}w(W,404,{ok:!1,error:"not_found"})}catch(J){V(`server error: ${J.message}`),w(W,500,{ok:!1,error:"internal"})}}),M0=2,E=0;function m$(){E+=1,f.listen(D,"127.0.0.1")}f.on("listening",()=>{e$(),V(`petdex sidecar listening on http://127.0.0.1:${D}`)});f.on("error",($)=>{if($.code!=="EADDRINUSE"){V(`server.error: ${$.message}`),process.exit(1);return}H0($)});async function H0($){if(E>=M0){V(`server.error: ${$.message}; gave up after ${E} attempts`),process.exit(1);return}let W=null;try{let Z=await fetch(`http://127.0.0.1:${D}/whoami`,{signal:AbortSignal.timeout(1000)});if(Z.ok)W=await Z.json()}catch{V(`server.error: ${$.message}; /whoami probe failed, yielding`),process.exit(1);return}if(!W||typeof W.pid!=="number"){V(`server.error: ${$.message}; /whoami had no pid, yielding`),process.exit(1);return}let J=W.parentPid,z=!0;if(typeof J==="number"&&J>0)try{process.kill(J,0)}catch(Z){z=Z.code!=="ESRCH"}else{V(`server.error: ${$.message}; incumbent pid=${W.pid} has no parentPid, yielding`),process.exit(1);return}if(z){V(`server.error: ${$.message}; incumbent pid=${W.pid} parent=${J} is alive, yielding`),process.exit(1);return}V(`server.error: ${$.message}; killing orphan sidecar pid=${W.pid} (parent=${J} dead)`);try{process.kill(W.pid,"SIGTERM")}catch(Z){V(`failed to SIGTERM orphan: ${Z.message}`),process.exit(1);return}if(!await O0(D,5000)){V(`port :${D} still busy 5s after killing orphan, giving up`),process.exit(1);return}V(`orphan evicted, retrying listen (attempt ${E+1})`),m$()}async function O0($,W){let J=Date.now()+W;while(Date.now()<J){if(await new Promise((Q)=>{let Z=a.default.createServer().once("error",()=>Q(!1)).once("listening",()=>Z.close(()=>Q(!0)));Z.listen($,"127.0.0.1")}))return!0;await new Promise((Q)=>setTimeout(Q,200))}return!1}m$();var b$=60000;function W$($){if(j){V(`shutdown(${$}) ignored: handoff in progress`);return}if(c){V(`sidecar received ${$}; update in progress`);let W=Date.now(),J=setTimeout(()=>{V(`update still running after ${b$}ms; forcing exit`);let z=y();if(z.status==="running")L({...z,status:"error",message:"Sidecar shut down before update finished. Re-launch Petdex and try again.",checkedAt:Date.now()});f.close(()=>process.exit(0)),setTimeout(()=>process.exit(0),1000).unref()},b$);if(!U){J.unref();return}U.finally(()=>{if(clearTimeout(J),j)return;V(`update settled after ${Date.now()-W}ms; shutting down`),f.close(()=>process.exit(0)),setTimeout(()=>process.exit(0),1000).unref()});return}V(`sidecar received ${$}, shutting down`),f.close(()=>process.exit(0)),setTimeout(()=>process.exit(0),1000).unref()}process.on("SIGTERM",()=>W$("SIGTERM"));process.on("SIGINT",()=>W$("SIGINT"));var L0=setTimeout(()=>{D$(),setInterval(()=>void D$(),a$).unref()},r$);L0.unref();var m=Number(process.env.PETDEX_PARENT_PID);if(Number.isFinite(m)&&m>0){V(`sidecar watching parent pid ${m}`);let $=setInterval(()=>{try{process.kill(m,0)}catch{if(c||I)return;V(`parent ${m} gone, exiting`),clearInterval($),W$("parent-gone")}},2000);$.unref()}
